Move real development code out of tests
