+++ dependency-inversion-ios/DependencyInversion/Fruit.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct Fruit {
+    let name: String
+}
